Refactor dashboard census service boundary
