+++ components/admin/animalSpecies/createViewModel.ts
@@ -1,4 +1,4 @@
-import { AnimalSpecies } from "lib/services/animalSpecies";
+import { AnimalSpecies } from "lib/services/census";
 import { SaveResult } from "lib/services/interface";
 import { AnimalSpeciesViewModel } from "./animalSpeciesViewModel";
