--- lib/services/animalSpecies/animalSpecies.ts
@@ -1,15 +1,0 @@
-export type AnimalSpecies = {
-  id: string;
-  code: string;
-  name: string;
-  active: boolean;
-  sortOrder: number;
-};
-
-export type AnimalCensusFact = {
-  speciesId: string;
-  speciesCode: string;
-  speciesName: string;
-  animalQuantity: number;
-  householdQuantity: number;
-};
