--- lib/services/animalSpecies/index.ts
@@ -1,3 +1,0 @@
-export { AnimalSpeciesService } from "./animalSpeciesService";
-export type { IAnimalSpeciesService } from "./animalSpeciesService";
-export type { AnimalSpecies, AnimalCensusFact } from "./animalSpecies";
