+++ lib/services/census/animalSpecies.ts
@@ -1,0 +1,7 @@
+export type AnimalSpecies = {
+  id: string;
+  code: string;
+  name: string;
+  active: boolean;
+  sortOrder: number;
+};
