+++ lib/services/census/index.ts
@@ -1,0 +1,17 @@
+export { AnimalSpeciesService } from "./animalSpeciesService";
+export type { IAnimalSpeciesService } from "./animalSpeciesService";
+export type { AnimalSpecies } from "./animalSpecies";
+
+export { CensusDefinitionService } from "./definitionService";
+export type { ICensusDefinitionService } from "./definitionService";
+export type {
+  CensusDefinition,
+  CensusDefinitionAdminState,
+  CensusDefinitionVersion,
+  CensusKind,
+  CensusSchema,
+} from "./definition";
+
+export { CensusSnapshotService } from "./snapshotService";
+export type { ICensusSnapshotService } from "./snapshotService";
+export type { AnimalCensusFact, VillageCensusSnapshot } from "./census";
