+++ lib/services/census/snapshot.graphql
@@ -1,0 +1,19 @@
+query LatestVillageCensus($villageId: Int!) {
+  latestVillageCensus(villageId: $villageId) {
+    id
+    censusDate
+    submittedAt
+    reporter {
+      username
+    }
+    facts {
+      species {
+        id
+        code
+        name
+      }
+      animalQuantity
+      householdQuantity
+    }
+  }
+}
