--- lib/services/censusDefinition/index.ts
@@ -1,9 +1,0 @@
-export { CensusDefinitionService } from "./censusDefinitionService";
-export type { ICensusDefinitionService } from "./censusDefinitionService";
-export type {
-  CensusDefinition,
-  CensusDefinitionAdminState,
-  CensusDefinitionVersion,
-  CensusKind,
-  CensusSchema,
-} from "./censusDefinition";
