+++ lib/services/provider.ts
@@ -24,8 +24,11 @@ import { ForgotPasswordService } from "./forgotPassword/forgotPasswordService";
 import { OutbreakPlanService } from "./outbreakPlan";
 import { PlaceService } from "./place";
 import { VillageService } from "./village";
-import { AnimalSpeciesService } from "./animalSpecies";
-import { CensusDefinitionService } from "./censusDefinition";
+import {
+  AnimalSpeciesService,
+  CensusDefinitionService,
+  CensusSnapshotService,
+} from "./census";
 import { ConfigurationService } from "./configuration";
 import { OutbreakService } from "./outbreak/outbreakService";
 import { ObservationDefinitionService } from "./observationDefinition/observationDefinitionService";
@@ -57,6 +60,7 @@ export interface IServiceProvider {
   get outbreakPlanService(): OutbreakPlanService;
   get placeService(): PlaceService;
   get villageService(): VillageService;
+  get censusSnapshotService(): CensusSnapshotService;
   get animalSpeciesService(): AnimalSpeciesService;
   get censusDefinitionService(): CensusDefinitionService;
   get configurationService(): ConfigurationService;
@@ -93,6 +97,7 @@ export class ServicesProvider implements IServiceProvider {
   outbreakPlanService: OutbreakPlanService;
   placeService: PlaceService;
   villageService: VillageService;
+  censusSnapshotService: CensusSnapshotService;
   animalSpeciesService: AnimalSpeciesService;
   censusDefinitionService: CensusDefinitionService;
   configurationService: ConfigurationService;
@@ -127,6 +132,7 @@ export class ServicesProvider implements IServiceProvider {
     this.outbreakPlanService = new OutbreakPlanService(client);
     this.placeService = new PlaceService(client);
     this.villageService = new VillageService(client);
+    this.censusSnapshotService = new CensusSnapshotService(client);
     this.animalSpeciesService = new AnimalSpeciesService(client);
     this.censusDefinitionService = new CensusDefinitionService(client);
     this.configurationService = new ConfigurationService(client);
