+++ lib/services/village/index.ts
@@ -1,3 +1,3 @@
 export { VillageService } from "./villageService";
 export type { IVillageService } from "./villageService";
-export type { Village, VillageCensusSnapshot } from "./village";
+export type { Village } from "./village";
